Improve license mangt
