+++ Website/package.json
@@ -8,9 +8,9 @@
   "author": "Der_Googler",
   "license": "ISC",
   "scripts": {
-    "dev": "npm run licenses && webpack --mode=development",
-    "prod": "npm run licenses && webpack --mode=production",
-    "licenses": "node license > ./src/utils/licenses.json"
+    "dev": "bun run licenses && webpack --mode=development",
+    "prod": "bun run licenses && webpack --mode=production",
+    "licenses": "bun license.ts > ./src/utils/licenses.json"
   },
   "typings": "types/onsenui.d.ts",
   "dependencies": {
@@ -20,28 +20,20 @@
     "@mui/icons-material": "^5.8.0",
     "@mui/material": "^5.8.0",
     "axios": "^0.27.2",
-    "bootstrap": "^5.1.3",
-    "findbyview": "^1.0.2",
+    "bota64": "^1.0.2",
     "googlers-tools": "^1.1.7",
-    "gradle-to-js": "^2.0.1",
-    "highlight.js": "^11.5.1",
     "jss": "^10.9.0",
     "jss-preset-default": "^10.9.0",
     "localized-strings": "^0.2.4",
-    "markdown-to-jsx": "^7.1.7",
     "marked-react": "^1.1.2",
     "material-icons": "^1.10.8",
     "object-assign": "^4.1.1",
     "onsenui": "^2.12.1",
     "react": "^18.2.0",
-    "react-bootstrap": "^2.3.1",
     "react-device-detect": "^2.2.2",
     "react-dom": "^18.2.0",
     "react-onsenui": "^1.12.0",
-    "react-syntax-highlighter": "^15.5.0",
-    "react-toastify": "^9.0.1",
-    "remark-shortcodes": "^0.3.1",
-    "visit-values": "^2.0.0"
+    "react-syntax-highlighter": "^15.5.0"
   },
   "devDependencies": {
     "@babel/core": "^7.16.5",
